requirements.txt
flask
redis
beautifulsoup4
pytest-playwright
